+++ app/src/main/java/org/simple/clinic/main/TheActivity.kt
@@ -28,8 +28,8 @@ import org.simple.clinic.home.patients.LoggedOutOnOtherDeviceDialog
 import org.simple.clinic.login.applock.AppLockConfig
 import org.simple.clinic.login.applock.AppLockScreenKey
 import org.simple.clinic.mobius.MobiusDelegate
+import org.simple.clinic.navigation.v2.History
 import org.simple.clinic.navigation.v2.Router
-import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.navigation.v2.compat.wrap
 import org.simple.clinic.registerorlogin.AuthenticationActivity
 import org.simple.clinic.remoteconfig.UpdateRemoteConfigWorker
@@ -299,7 +299,7 @@ class TheActivity : AppCompatActivity(), TheActivityUi {
     router.clearHistoryAndPush(AccessDeniedScreenKey(fullName))
   }
 
-  override fun showInitialScreen(screenKey: ScreenKey) {
-    router.clearHistoryAndPush(screenKey)
+  override fun setCurrentScreenHistory(newHistory: History) {
+    router.replaceHistory(newHistory)
   }
 }